+++ context/app/routes_main.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 
 from flask import (Blueprint, render_template, abort, current_app,
-                   session, flash, get_flashed_messages,
+                   session, flash, get_flashed_messages, request,
                    redirect, url_for)
 
 from yaml import safe_load as load_yaml
@@ -43,7 +43,6 @@ def details(type, uuid):
     if 'nexus_token' not in session:
         abort(403)
     client = _get_client()
-
     entity = client.get_entity(uuid)
     actual_type = entity['entity_type'].lower()
     if type != actual_type:
@@ -74,7 +73,7 @@ def details(type, uuid):
         'flashed_messages': flashed_messages,
         'entity': entity,
         'provenance': provenance,
-        'vitessce_conf': client.get_vitessce_conf(),
+        'vitessce_conf': client.get_vitessce_conf(entity)
     })
     return render_template(
         template,
@@ -100,12 +99,18 @@ def details_ext(type, uuid, ext):
 
 @blueprint.route('/search')
 def search():
-    core_props = {'endpoints':
-    {'esEndpoint': current_app.config['ELASTICSEARCH_ENDPOINT'], 'assetsEndpoint': current_app.config['ASSETS_ENDPOINT']}}
+    entity_type = request.args.get('entity_type[0]')
+    title = f'{entity_type}s' if entity_type else 'Search'
+    core_props = {
+        'endpoints': {'esEndpoint': current_app.config['ELASTICSEARCH_ENDPOINT'],
+                      'assetsEndpoint': current_app.config['ASSETS_ENDPOINT']},
+        'title': title
+    }
     if 'nexus_token' not in session:
         abort(403)
     return render_template(
         'pages/base_react.html',
-        types = types,
-        flask_data = core_props
+        title=title,
+        types=types,
+        flask_data=core_props
     )